requirements.txt
scikit-learn
nose
sphinx
xgboost
hyperopt
django
sphinx-rtd-theme

mysql-connector-python
pandas
numpy
scipy